# [11.x] Optimize commands registry