--- a/src/Illuminate/Foundation/Console/OptimizeCommand.php
+++ b/src/Illuminate/Foundation/Console/OptimizeCommand.php
@@ -3,6 +3,7 @@
 namespace Illuminate\Foundation\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\ServiceProvider;
 use Symfony\Component\Console\Attribute\AsCommand;
 
 #[AsCommand(name: 'optimize')]
@@ -31,13 +32,21 @@ class OptimizeCommand extends Command
     {
         $this->components->info('Caching framework bootstrap, configuration, and metadata.');
 
-        collect([
-            'config' => fn () => $this->callSilent('config:cache') == 0,
-            'events' => fn () => $this->callSilent('event:cache') == 0,
-            'routes' => fn () => $this->callSilent('route:cache') == 0,
-            'views' => fn () => $this->callSilent('view:cache') == 0,
-        ])->each(fn ($task, $description) => $this->components->task($description, $task));
+        foreach ($this->getOptimizeTasks() as $description => $command) {
+            $this->components->task($description, fn () => $this->callSilently($command) == 0);
+        }
 
         $this->newLine();
+    }
+
+    protected function getOptimizeTasks(): array
+    {
+        return [
+            'config' => 'config:cache',
+            'events' => 'event:cache',
+            'routes' => 'route:cache',
+            'views' => 'view:cache',
+            ...ServiceProvider::$optimizingCommands,
+        ];
     }
 }