@ollieread 2024-09-26
Nice idea @erikgaal! I'd be inclined to have the third-party packages have the commands register as optimisation and optimisation clearing commands, similar to how you can register publishable assets.

```php
$this->registerOptimizationCommands(
    optimization: 'icons:cache', 
    optimizationClearing: 'icons:clear'
);
```

That way it's not silent, and then it opens you up to an interactive command, something like

```bash
php artisan optimize --select
```

That lets you select specifically which things you want to optimise.

That's my thinking on this, for whatever it's worth. Either way, good work so far and good luck with the PR.