@taylorotwell 2024-09-27
I would actually prefer just having the events like you originally had. 🫠