@erikgaal 2024-09-27
That would mean

- not having any output (the original implementation), or;
- leaking the command `$components` through the event and having to use `Artisan::callSilently` manually, or;
- providing something on the `Optimizing` and `OptimizeClearing` events that abstracts away the command `$components`. `$event->task(key: 'my-package', fn () => Artisan::callSilently('my-package:cache'))` 

Let me know what option you prefer and I'll get it done.