@shaedrich 2024-09-28
> ```
> $ php artisan optimize
> 
>    INFO  Caching framework bootstrap, configuration, and metadata.  
> 
>   config .................................................................... 39.29ms DONE
>   events ..................................................................... 2.28ms DONE
>   routes .................................................................... 36.46ms DONE
>   views .................................................................... 647.20ms DONE
>   my package ................................................................. 0.23ms DONE
> ```

I would actually prefer something like this:
```
$ php artisan optimize

   INFO  Caching framework bootstrap, configuration, and metadata.  

  config .................................................................... 39.29ms DONE
  events ..................................................................... 2.28ms DONE
  routes .................................................................... 36.46ms DONE
  views .................................................................... 647.20ms DONE
  my package:config .......................................................... 0.13ms DONE
  my package:routes .......................................................... 0.09ms DONE
```
or

```
$ php artisan optimize

   INFO  Caching framework bootstrap, configuration, and metadata.  

  config .................................................................... 39.29ms DONE
  events ..................................................................... 2.28ms DONE
  routes .................................................................... 36.46ms DONE
  views .................................................................... 647.20ms DONE
  my package ................................................................. 0.22ms DONE
      config ................................................................. 0.13ms DONE
      routes ................................................................. 0.09ms DONE
```

Alternatively, couldn't classes just implement something like a `Cachable` contract and all classes that have it, are collected for caching?